+++ src/flames/cloneFlames.cpp
@@ -22,7 +22,7 @@ bool CloneFlame::build()
   m_ctrlPointsSave = m_ctrlPoints;
   for(uint i=0; i < (m_maxParticles + m_nbFixedPoints) * (m_nbSkeletons ) * 3; i++)
     {
-      *m_ctrlPoints++ += rand()/(50*(double)RAND_MAX) - .025;
+      *m_ctrlPoints++ += rand()/(50*(float)RAND_MAX) - .025;
     }
   /* On recopie les m_uorder squelettes pour fermer la NURBS */
   GLfloat *startCtrlPoints = m_ctrlPointsSave;
